+++ accounting/src/apideck_accounting_unify/accounting.py
@@ -2,6 +2,7 @@
 
 from .basesdk import BaseSDK
 from .sdkconfiguration import SDKConfiguration
+from apideck_accounting_unify.agedcreditors_sdk import AgedCreditorsSDK
 from apideck_accounting_unify.ageddebtors_sdk import AgedDebtorsSDK
 from apideck_accounting_unify.attachments import Attachments
 from apideck_accounting_unify.balancesheet_sdk import BalanceSheetSDK
@@ -48,6 +49,7 @@ class Accounting(BaseSDK):
     tracking_categories: TrackingCategories
     bill_payments: BillPayments
     expenses: Expenses
+    aged_creditors: AgedCreditorsSDK
     aged_debtors: AgedDebtorsSDK
 
     def __init__(self, sdk_config: SDKConfiguration) -> None:
@@ -77,4 +79,5 @@ class Accounting(BaseSDK):
         self.tracking_categories = TrackingCategories(self.sdk_configuration)
         self.bill_payments = BillPayments(self.sdk_configuration)
         self.expenses = Expenses(self.sdk_configuration)
+        self.aged_creditors = AgedCreditorsSDK(self.sdk_configuration)
         self.aged_debtors = AgedDebtorsSDK(self.sdk_configuration)