+++ accounting/src/apideck_accounting_unify/attachments.py
@@ -6,8 +6,9 @@ from apideck_accounting_unify._hooks import HookContext
 from apideck_accounting_unify.types import OptionalNullable, UNSET
 from apideck_accounting_unify.utils import get_security_from_env
 from enum import Enum
+import io
 from jsonpath import JSONPath
-from typing import Any, Dict, Mapping, Optional
+from typing import Any, Dict, IO, Mapping, Optional, Union
 
 
 class DownloadAcceptEnum(str, Enum):
@@ -383,6 +384,340 @@ class Attachments(BaseSDK):
                 ),
                 http_meta=models.HTTPMetadata(request=req, response=http_res),
                 next=next_func,
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = await utils.stream_to_text_async(http_res)
+        raise models.APIError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )
+
+    def upload(
+        self,
+        *,
+        reference_type: models.AttachmentReferenceType,
+        reference_id: str,
+        request_body: Union[bytes, IO[bytes], io.BufferedReader],
+        raw: Optional[bool] = False,
+        x_apideck_metadata: Optional[str] = None,
+        consumer_id: Optional[str] = None,
+        app_id: Optional[str] = None,
+        service_id: Optional[str] = None,
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+        http_headers: Optional[Mapping[str, str]] = None,
+    ) -> models.AccountingAttachmentsUploadResponse:
+        r"""Upload attachment
+
+        Upload attachment
+
+        :param reference_type: The reference type of the document.
+        :param reference_id: The reference id of the object to retrieve.
+        :param request_body:
+        :param raw: Include raw response. Mostly used for debugging purposes
+        :param x_apideck_metadata: Metadata to attach to the attachment file (JSON string)
+        :param consumer_id: ID of the consumer which you want to get or push data from
+        :param app_id: The ID of your Unify application
+        :param service_id: Provide the service id you want to call (e.g., pipedrive). Only needed when a consumer has activated multiple integrations for a Unified API.
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        :param http_headers: Additional headers to set or replace on requests.
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+        else:
+            base_url = models.ACCOUNTING_ATTACHMENTS_UPLOAD_OP_SERVERS[0]
+
+        request = models.AccountingAttachmentsUploadRequest(
+            reference_type=reference_type,
+            reference_id=reference_id,
+            raw=raw,
+            x_apideck_metadata=x_apideck_metadata,
+            consumer_id=consumer_id,
+            app_id=app_id,
+            service_id=service_id,
+            request_body=request_body,
+        )
+
+        req = self._build_request(
+            method="POST",
+            path="/accounting/attachments/{reference_type}/{reference_id}",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=True,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            http_headers=http_headers,
+            _globals=models.AccountingAttachmentsUploadGlobals(
+                consumer_id=self.sdk_configuration.globals.consumer_id,
+                app_id=self.sdk_configuration.globals.app_id,
+            ),
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.request_body,
+                False,
+                False,
+                "raw",
+                Union[bytes, IO[bytes], io.BufferedReader],
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["5XX"])
+
+        http_res = self.do_request(
+            hook_ctx=HookContext(
+                base_url=base_url or "",
+                operation_id="accounting.attachmentsUpload",
+                oauth2_scopes=[],
+                security_source=get_security_from_env(
+                    self.sdk_configuration.security, models.Security
+                ),
+            ),
+            request=req,
+            error_status_codes=["400", "401", "402", "404", "422", "4XX", "5XX"],
+            retry_config=retry_config,
+        )
+
+        response_data: Any = None
+        if utils.match_response(http_res, "201", "application/json"):
+            return models.AccountingAttachmentsUploadResponse(
+                create_attachment_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.CreateAttachmentResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+        if utils.match_response(http_res, "400", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.BadRequestResponseData
+            )
+            raise models.BadRequestResponse(data=response_data)
+        if utils.match_response(http_res, "401", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnauthorizedResponseData
+            )
+            raise models.UnauthorizedResponse(data=response_data)
+        if utils.match_response(http_res, "402", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.PaymentRequiredResponseData
+            )
+            raise models.PaymentRequiredResponse(data=response_data)
+        if utils.match_response(http_res, "404", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.NotFoundResponseData
+            )
+            raise models.NotFoundResponse(data=response_data)
+        if utils.match_response(http_res, "422", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnprocessableResponseData
+            )
+            raise models.UnprocessableResponse(data=response_data)
+        if utils.match_response(http_res, "4XX", "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
+            http_res_text = utils.stream_to_text(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "default", "application/json"):
+            return models.AccountingAttachmentsUploadResponse(
+                unexpected_error_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.UnexpectedErrorResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+
+        content_type = http_res.headers.get("Content-Type")
+        http_res_text = utils.stream_to_text(http_res)
+        raise models.APIError(
+            f"Unexpected response received (code: {http_res.status_code}, type: {content_type})",
+            http_res.status_code,
+            http_res_text,
+            http_res,
+        )
+
+    async def upload_async(
+        self,
+        *,
+        reference_type: models.AttachmentReferenceType,
+        reference_id: str,
+        request_body: Union[bytes, IO[bytes], io.BufferedReader],
+        raw: Optional[bool] = False,
+        x_apideck_metadata: Optional[str] = None,
+        consumer_id: Optional[str] = None,
+        app_id: Optional[str] = None,
+        service_id: Optional[str] = None,
+        retries: OptionalNullable[utils.RetryConfig] = UNSET,
+        server_url: Optional[str] = None,
+        timeout_ms: Optional[int] = None,
+        http_headers: Optional[Mapping[str, str]] = None,
+    ) -> models.AccountingAttachmentsUploadResponse:
+        r"""Upload attachment
+
+        Upload attachment
+
+        :param reference_type: The reference type of the document.
+        :param reference_id: The reference id of the object to retrieve.
+        :param request_body:
+        :param raw: Include raw response. Mostly used for debugging purposes
+        :param x_apideck_metadata: Metadata to attach to the attachment file (JSON string)
+        :param consumer_id: ID of the consumer which you want to get or push data from
+        :param app_id: The ID of your Unify application
+        :param service_id: Provide the service id you want to call (e.g., pipedrive). Only needed when a consumer has activated multiple integrations for a Unified API.
+        :param retries: Override the default retry configuration for this method
+        :param server_url: Override the default server URL for this method
+        :param timeout_ms: Override the default request timeout configuration for this method in milliseconds
+        :param http_headers: Additional headers to set or replace on requests.
+        """
+        base_url = None
+        url_variables = None
+        if timeout_ms is None:
+            timeout_ms = self.sdk_configuration.timeout_ms
+
+        if server_url is not None:
+            base_url = server_url
+        else:
+            base_url = models.ACCOUNTING_ATTACHMENTS_UPLOAD_OP_SERVERS[0]
+
+        request = models.AccountingAttachmentsUploadRequest(
+            reference_type=reference_type,
+            reference_id=reference_id,
+            raw=raw,
+            x_apideck_metadata=x_apideck_metadata,
+            consumer_id=consumer_id,
+            app_id=app_id,
+            service_id=service_id,
+            request_body=request_body,
+        )
+
+        req = self._build_request_async(
+            method="POST",
+            path="/accounting/attachments/{reference_type}/{reference_id}",
+            base_url=base_url,
+            url_variables=url_variables,
+            request=request,
+            request_body_required=True,
+            request_has_path_params=True,
+            request_has_query_params=True,
+            user_agent_header="user-agent",
+            accept_header_value="application/json",
+            http_headers=http_headers,
+            _globals=models.AccountingAttachmentsUploadGlobals(
+                consumer_id=self.sdk_configuration.globals.consumer_id,
+                app_id=self.sdk_configuration.globals.app_id,
+            ),
+            security=self.sdk_configuration.security,
+            get_serialized_body=lambda: utils.serialize_request_body(
+                request.request_body,
+                False,
+                False,
+                "raw",
+                Union[bytes, IO[bytes], io.BufferedReader],
+            ),
+            timeout_ms=timeout_ms,
+        )
+
+        if retries == UNSET:
+            if self.sdk_configuration.retry_config is not UNSET:
+                retries = self.sdk_configuration.retry_config
+            else:
+                retries = utils.RetryConfig(
+                    "backoff", utils.BackoffStrategy(500, 60000, 1.5, 3600000), True
+                )
+
+        retry_config = None
+        if isinstance(retries, utils.RetryConfig):
+            retry_config = (retries, ["5XX"])
+
+        http_res = await self.do_request_async(
+            hook_ctx=HookContext(
+                base_url=base_url or "",
+                operation_id="accounting.attachmentsUpload",
+                oauth2_scopes=[],
+                security_source=get_security_from_env(
+                    self.sdk_configuration.security, models.Security
+                ),
+            ),
+            request=req,
+            error_status_codes=["400", "401", "402", "404", "422", "4XX", "5XX"],
+            retry_config=retry_config,
+        )
+
+        response_data: Any = None
+        if utils.match_response(http_res, "201", "application/json"):
+            return models.AccountingAttachmentsUploadResponse(
+                create_attachment_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.CreateAttachmentResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
+            )
+        if utils.match_response(http_res, "400", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.BadRequestResponseData
+            )
+            raise models.BadRequestResponse(data=response_data)
+        if utils.match_response(http_res, "401", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnauthorizedResponseData
+            )
+            raise models.UnauthorizedResponse(data=response_data)
+        if utils.match_response(http_res, "402", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.PaymentRequiredResponseData
+            )
+            raise models.PaymentRequiredResponse(data=response_data)
+        if utils.match_response(http_res, "404", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.NotFoundResponseData
+            )
+            raise models.NotFoundResponse(data=response_data)
+        if utils.match_response(http_res, "422", "application/json"):
+            response_data = utils.unmarshal_json(
+                http_res.text, models.UnprocessableResponseData
+            )
+            raise models.UnprocessableResponse(data=response_data)
+        if utils.match_response(http_res, "4XX", "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "5XX", "*"):
+            http_res_text = await utils.stream_to_text_async(http_res)
+            raise models.APIError(
+                "API error occurred", http_res.status_code, http_res_text, http_res
+            )
+        if utils.match_response(http_res, "default", "application/json"):
+            return models.AccountingAttachmentsUploadResponse(
+                unexpected_error_response=utils.unmarshal_json(
+                    http_res.text, Optional[models.UnexpectedErrorResponse]
+                ),
+                http_meta=models.HTTPMetadata(request=req, response=http_res),
             )
 
         content_type = http_res.headers.get("Content-Type")