+++ accounting/src/apideck_accounting_unify/models/accounting_invoiceitemsoneop.py
@@ -133,7 +133,7 @@ class AccountingInvoiceItemsOneRequest(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)