+++ accounting/src/apideck_accounting_unify/models/accounting_ledgeraccountsallop.py
@@ -162,7 +162,7 @@ class AccountingLedgerAccountsAllRequest(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)