+++ accounting/src/apideck_accounting_unify/models/accountingdepartment.py
@@ -123,7 +123,7 @@ class AccountingDepartment(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)
@@ -193,7 +193,7 @@ class AccountingDepartmentInput(BaseModel):
 
         m = {}
 
-        for n, f in self.model_fields.items():
+        for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
             serialized.pop(k, None)