+++ accounting/src/apideck_accounting_unify/purchaseorders.py
@@ -456,6 +456,9 @@ class PurchaseOrders(BaseSDK):
                 List[Nullable[models.LinkedTrackingCategoryTypedDict]],
             ]
         ] = UNSET,
+        custom_fields: Optional[
+            Union[List[models.CustomField], List[models.CustomFieldTypedDict]]
+        ] = None,
         row_version: OptionalNullable[str] = UNSET,
         pass_through: Optional[
             Union[List[models.PassThroughBody], List[models.PassThroughBodyTypedDict]]
@@ -500,6 +503,7 @@ class PurchaseOrders(BaseSDK):
         :param channel: The channel through which the transaction is processed.
         :param memo: Message for the supplier. This text appears on the Purchase Order.
         :param tracking_categories: A list of linked tracking categories.
+        :param custom_fields:
         :param row_version: A binary value used to detect updates to a object and prevent data conflicts. It is incremented each time an update is made to the object.
         :param pass_through: The pass_through property allows passing service-specific, custom data or structured modifications in request body when creating or updating resources.
         :param retries: Override the default retry configuration for this method
@@ -562,6 +566,9 @@ class PurchaseOrders(BaseSDK):
                 tracking_categories=utils.get_pydantic_model(
                     tracking_categories,
                     OptionalNullable[List[Nullable[models.LinkedTrackingCategory]]],
+                ),
+                custom_fields=utils.get_pydantic_model(
+                    custom_fields, Optional[List[models.CustomField]]
                 ),
                 row_version=row_version,
                 pass_through=utils.get_pydantic_model(
@@ -734,6 +741,9 @@ class PurchaseOrders(BaseSDK):
                 List[Nullable[models.LinkedTrackingCategoryTypedDict]],
             ]
         ] = UNSET,
+        custom_fields: Optional[
+            Union[List[models.CustomField], List[models.CustomFieldTypedDict]]
+        ] = None,
         row_version: OptionalNullable[str] = UNSET,
         pass_through: Optional[
             Union[List[models.PassThroughBody], List[models.PassThroughBodyTypedDict]]
@@ -778,6 +788,7 @@ class PurchaseOrders(BaseSDK):
         :param channel: The channel through which the transaction is processed.
         :param memo: Message for the supplier. This text appears on the Purchase Order.
         :param tracking_categories: A list of linked tracking categories.
+        :param custom_fields:
         :param row_version: A binary value used to detect updates to a object and prevent data conflicts. It is incremented each time an update is made to the object.
         :param pass_through: The pass_through property allows passing service-specific, custom data or structured modifications in request body when creating or updating resources.
         :param retries: Override the default retry configuration for this method
@@ -840,6 +851,9 @@ class PurchaseOrders(BaseSDK):
                 tracking_categories=utils.get_pydantic_model(
                     tracking_categories,
                     OptionalNullable[List[Nullable[models.LinkedTrackingCategory]]],
+                ),
+                custom_fields=utils.get_pydantic_model(
+                    custom_fields, Optional[List[models.CustomField]]
                 ),
                 row_version=row_version,
                 pass_through=utils.get_pydantic_model(
@@ -1315,6 +1329,9 @@ class PurchaseOrders(BaseSDK):
                 List[Nullable[models.LinkedTrackingCategoryTypedDict]],
             ]
         ] = UNSET,
+        custom_fields: Optional[
+            Union[List[models.CustomField], List[models.CustomFieldTypedDict]]
+        ] = None,
         row_version: OptionalNullable[str] = UNSET,
         pass_through: Optional[
             Union[List[models.PassThroughBody], List[models.PassThroughBodyTypedDict]]
@@ -1360,6 +1377,7 @@ class PurchaseOrders(BaseSDK):
         :param channel: The channel through which the transaction is processed.
         :param memo: Message for the supplier. This text appears on the Purchase Order.
         :param tracking_categories: A list of linked tracking categories.
+        :param custom_fields:
         :param row_version: A binary value used to detect updates to a object and prevent data conflicts. It is incremented each time an update is made to the object.
         :param pass_through: The pass_through property allows passing service-specific, custom data or structured modifications in request body when creating or updating resources.
         :param retries: Override the default retry configuration for this method
@@ -1423,6 +1441,9 @@ class PurchaseOrders(BaseSDK):
                 tracking_categories=utils.get_pydantic_model(
                     tracking_categories,
                     OptionalNullable[List[Nullable[models.LinkedTrackingCategory]]],
+                ),
+                custom_fields=utils.get_pydantic_model(
+                    custom_fields, Optional[List[models.CustomField]]
                 ),
                 row_version=row_version,
                 pass_through=utils.get_pydantic_model(
@@ -1596,6 +1617,9 @@ class PurchaseOrders(BaseSDK):
                 List[Nullable[models.LinkedTrackingCategoryTypedDict]],
             ]
         ] = UNSET,
+        custom_fields: Optional[
+            Union[List[models.CustomField], List[models.CustomFieldTypedDict]]
+        ] = None,
         row_version: OptionalNullable[str] = UNSET,
         pass_through: Optional[
             Union[List[models.PassThroughBody], List[models.PassThroughBodyTypedDict]]
@@ -1641,6 +1665,7 @@ class PurchaseOrders(BaseSDK):
         :param channel: The channel through which the transaction is processed.
         :param memo: Message for the supplier. This text appears on the Purchase Order.
         :param tracking_categories: A list of linked tracking categories.
+        :param custom_fields:
         :param row_version: A binary value used to detect updates to a object and prevent data conflicts. It is incremented each time an update is made to the object.
         :param pass_through: The pass_through property allows passing service-specific, custom data or structured modifications in request body when creating or updating resources.
         :param retries: Override the default retry configuration for this method
@@ -1704,6 +1729,9 @@ class PurchaseOrders(BaseSDK):
                 tracking_categories=utils.get_pydantic_model(
                     tracking_categories,
                     OptionalNullable[List[Nullable[models.LinkedTrackingCategory]]],
+                ),
+                custom_fields=utils.get_pydantic_model(
+                    custom_fields, Optional[List[models.CustomField]]
                 ),
                 row_version=row_version,
                 pass_through=utils.get_pydantic_model(